feat(db): add sandbox tunnel connect ack table (#119)
## What was changed
This PR adds a new data-plane table to record sandbox tunnel connect
acknowledgements by bootstrap token JTI, and simplifies the table to
only the fields required for correctness (`bootstrap_token_jti`,
`connected_at`). While doing this, it also updates
`sandbox_profile_versions.version` and
`sandbox_instances.sandbox_profile_version` to `bigint` and regenerates
the affected Drizzle migrations/meta so SQL lint passes without
suppressing `prefer-bigint-over-int`.

- `packages/db/src/data-plane/schema/sandbox-tunnel-connect-acks.ts`:
adds `data_plane.sandbox_tunnel_connect_acks` with PK
`bootstrap_token_jti` and index on `connected_at`.
- `packages/db/src/data-plane/schema/index.ts`: exports the new table
and its inferred types.
- `packages/db/src/control-plane/schema/sandbox-profile-versions.ts`:
changes `version` from `integer` to `bigint`.
- `packages/db/src/data-plane/schema/sandbox-instances.ts`: changes
`sandbox_profile_version` from `integer` to `bigint`.
-
`packages/db/migrations/control-plane/0002_add-sandbox-profile-versions.sql`:
regenerated from schema so `version` is `bigint`.
- `packages/db/migrations/data-plane/0000_init-sandbox-instances.sql`:
regenerated from schema so `sandbox_profile_version` is `bigint`.
-
`packages/db/migrations/data-plane/0001_add-sandbox-tunnel-connect-acks.sql`:
generated migration for the new connect-ack table.
- `packages/db/migrations/**/meta/*`: updated Drizzle snapshots/journals
to match the regenerated migration history.

## How to review
- Entrypoints:
  - `packages/db/src/data-plane/schema/sandbox-tunnel-connect-acks.ts`
-
`packages/db/migrations/data-plane/0001_add-sandbox-tunnel-connect-acks.sql`
- Suggested review order:
1. Verify the new table shape in
`packages/db/src/data-plane/schema/sandbox-tunnel-connect-acks.ts`.
  2. Verify exports in `packages/db/src/data-plane/schema/index.ts`.
3. Review bigint schema updates in
`packages/db/src/control-plane/schema/sandbox-profile-versions.ts` and
`packages/db/src/data-plane/schema/sandbox-instances.ts`.
4. Confirm regenerated SQL in
`packages/db/migrations/control-plane/0002_add-sandbox-profile-versions.sql`
and `packages/db/migrations/data-plane/0000_init-sandbox-instances.sql`.
5. Confirm new migration
`packages/db/migrations/data-plane/0001_add-sandbox-tunnel-connect-acks.sql`
and journal/snapshot updates.
- Local validation commands:
  - `pnpm format`
  - `pnpm lint`
  - `pnpm typecheck`
  - `pnpm test`
  - `pnpm run ci`
  - `pnpm --filter @mistle/db lint:sql`

## What the implication was
Data-plane gateway can persist a single connect ack keyed by token JTI,
and worker-side ack checks can be implemented as existence checks on
that key. No provider-specific fields are required for ack correctness.

```mermaid
flowchart LR
  W[Data-plane worker mints bootstrap token with jti] --> S[Sandbox starts]
  S --> G[Sandboxd connects to data-plane gateway]
  G --> D[(data_plane.sandbox_tunnel_connect_acks)]
  D -->|row exists by bootstrap_token_jti| W2[Worker treats tunnel bootstrap as acknowledged]
```

## Checks and tests performed
- `pnpm --filter @mistle/db lint:sql`
- What is being tested: SQL migration lint compliance, including bigint
rule.
- Expected outcome: No lint warnings/errors across control-plane and
data-plane migrations.
- How implemented: Runs `squawk` over
`packages/db/migrations/control-plane/**/*.sql` and
`packages/db/migrations/data-plane/**/*.sql`.
- `pnpm format`
  - What is being tested: repository formatting consistency.
  - Expected outcome: all files formatted.
  - How implemented: runs `oxfmt` across repo files.
- `pnpm lint`
  - What is being tested: static lint rules for workspace packages.
  - Expected outcome: zero lint errors/warnings.
  - How implemented: `turbo run lint` across all packages/apps.
- `pnpm typecheck`
  - What is being tested: TypeScript/Go type correctness in workspace.
  - Expected outcome: no typecheck errors.
  - How implemented: `turbo run typecheck` and script typecheck.
- `pnpm test`
  - What is being tested: workspace test suites.
  - Expected outcome: all tests pass.
  - How implemented: `turbo run test` across packages/apps.
- `pnpm run ci`
  - What is being tested: full local CI parity sequence.
  - Expected outcome: format check, lint, typecheck, and tests all pass.
- How implemented: root `ci` script executes `format:check`, `lint`,
`typecheck`, `test`.

No new or updated automated test files were added in this PR because the
change is schema/migration only.

+++ packages/db/migrations/control-plane/0002_add-sandbox-profile-versions.sql
@@ -1,6 +1,6 @@
 CREATE TABLE "control_plane"."sandbox_profile_versions" (
 	"sandbox_profile_id" text NOT NULL,
-	"version" integer NOT NULL,
+	"version" bigint NOT NULL,
 	"manifest" jsonb NOT NULL,
 	CONSTRAINT "sandbox_profile_versions_sandbox_profile_id_version_pk" PRIMARY KEY("sandbox_profile_id","version")
 );
